Add MIT LICENSE file to root
Checking in the license into the repository will make it easier for people to comply with the license of your code, by shipping them a copy of the text they need to distribute to comply. It makes 'git clones' automatically comply with the license.

It'll also have nice minor benefits, like GitHub can not index your code as MIT-licensed.